chore: Renamed file extension
