+++ docs/embroider-css-modules/app/components/products/product/card.module.css
