+++ docs/embroider-css-modules/app/components/products/product/image.gts
@@ -1,7 +1,7 @@
 import type { TOC } from '@ember/component/template-only';
 
 import config from '../../../config/environment';
-import styles from './image.css';
+import styles from './image.module.css';
 
 const isTestEnvironment = config.environment === 'test';
