+++ docs/embroider-css-modules/app/components/products/product/image.module.css
