+++ docs/embroider-css-modules/app/components/widgets/widget-1.module.css
