+++ docs/embroider-css-modules/app/components/widgets/widget-2.module.css
