+++ docs/embroider-css-modules/app/components/widgets/widget-2/stacked-chart.module.css
