+++ docs/embroider-css-modules/app/components/widgets/widget-3.module.css
