+++ docs/embroider-css-modules/app/components/widgets/widget-3/tour-schedule.gts
@@ -3,7 +3,7 @@ import { hash } from '@ember/helper';
 import { ContainerQuery, width } from 'ember-container-query';
 
 import type { Concert } from '../../../data/concert';
-import styles from './tour-schedule.css';
+import styles from './tour-schedule.module.css';
 import WidgetsWidget3TourScheduleResponsiveImage from './tour-schedule/responsive-image';
 
 interface WidgetsWidget3TourScheduleSignature {
