+++ docs/embroider-css-modules/app/components/widgets/widget-3/tour-schedule/responsive-image.gts
@@ -5,7 +5,7 @@ import { containerQuery, type Dimensions } from 'ember-container-query';
 
 import type { Image } from '../../../../data/concert';
 import { findBestFittingImage } from '../../../../utils/components/widgets/widget-3';
-import styles from './responsive-image.css';
+import styles from './responsive-image.module.css';
 
 interface WidgetsWidget3TourScheduleResponsiveImageSignature {
   Args: {
