+++ docs/embroider-css-modules/app/components/widgets/widget-4.module.css
