+++ docs/embroider-css-modules/app/components/widgets/widget-5.gts
@@ -3,7 +3,7 @@ import { hash } from '@ember/helper';
 import { ContainerQuery, height, width } from 'ember-container-query';
 
 import strictAnd from '../../helpers/strict-and';
-import styles from './widget-5.css';
+import styles from './widget-5.module.css';
 
 interface WidgetsWidget5Signature {}
