+++ docs/embroider-css-modules/app/components/widgets/widget-5.module.css
