+++ docs/embroider-css-modules/app/controllers/application.ts
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
 
-import styles from './application.css';
+import styles from './application.module.css';
 
 export default class ApplicationController extends Controller {
   styles = styles;
