+++ docs/embroider-css-modules/app/controllers/dashboard.ts
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
 
-import styles from './dashboard.css';
+import styles from './dashboard.module.css';
 
 export default class DashboardController extends Controller {
   styles = styles;
