+++ docs/embroider-css-modules/app/controllers/form.ts
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
 
-import styles from './form.css';
+import styles from './form.module.css';
 
 export default class FormController extends Controller {
   styles = styles;
