+++ docs/embroider-css-modules/app/controllers/index.ts
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
 
-import styles from './index.css';
+import styles from './index.module.css';
 
 export default class IndexController extends Controller {
   styles = styles;
