+++ docs/embroider-css-modules/app/controllers/not-found.ts
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
 
-import styles from './not-found.css';
+import styles from './not-found.module.css';
 
 export default class NotFoundController extends Controller {
   styles = styles;
