+++ docs/embroider-css-modules/tests/integration/components/ui/form/field-test.ts
@@ -1,6 +1,6 @@
 import type { TestContext } from '@ember/test-helpers';
 import { render } from '@ember/test-helpers';
-import styles from 'docs-app-for-embroider-css-modules/components/ui/form/field.css';
+import styles from 'docs-app-for-embroider-css-modules/components/ui/form/field.module.css';
 import { a11yAudit } from 'ember-a11y-testing/test-support';
 import { hbs } from 'ember-cli-htmlbars';
 import { module, test } from 'qunit';
